Refactor 404's into managers & drivers (bug 968519)
The goal is to move the responsibility of reference checks away from
controllers and into the underlying managers & drivers, which can
handle the task with equal or greater efficiency.

- Tenant references from create_user/update_user are NOT tested
  due to inconsistencies between backends
- Additional test coverage improvements

Also fixes bug 999209, bug 999608, bug 1006029, bug 1006055, bug 1006287,
bug 1006334, and bug 1006344.

+++ keystone/identity/backends/kvs.py
@@ -45,49 +45,66 @@ class Identity(kvs.Base, identity.Driver):
         in the list of tenants on the user.
 
         """
-        user_ref = self._get_user(user_id)
+        user_ref = None
         tenant_ref = None
-        metadata_ref = None
-        if (not user_ref
-                or not utils.check_password(password,
-                                            user_ref.get('password'))):
+        metadata_ref = {}
+
+        try:
+            user_ref = self._get_user(user_id)
+        except exception.UserNotFound:
             raise AssertionError('Invalid user / password')
 
-        tenants = self.get_tenants_for_user(user_id)
-        if tenant_id and tenant_id not in tenants:
-            raise AssertionError('Invalid tenant')
+        if not utils.check_password(password, user_ref.get('password')):
+            raise AssertionError('Invalid user / password')
 
-        tenant_ref = self.get_tenant(tenant_id)
-        if tenant_ref:
-            metadata_ref = self.get_metadata(user_id, tenant_id)
-        else:
-            metadata_ref = {}
+        if tenant_id is not None:
+            if tenant_id not in self.get_tenants_for_user(user_id):
+                raise AssertionError('Invalid tenant')
+
+            try:
+                tenant_ref = self.get_tenant(tenant_id)
+                metadata_ref = self.get_metadata(user_id, tenant_id)
+            except exception.TenantNotFound:
+                tenant_ref = None
+                metadata_ref = {}
+            except exception.MetadataNotFound:
+                metadata_ref = {}
+
         return (_filter_user(user_ref), tenant_ref, metadata_ref)
 
     def get_tenant(self, tenant_id):
-        tenant_ref = self.db.get('tenant-%s' % tenant_id)
-        return tenant_ref
+        try:
+            return self.db.get('tenant-%s' % tenant_id)
+        except exception.NotFound:
+            raise exception.TenantNotFound(tenant_id=tenant_id)
 
     def get_tenants(self):
         tenant_keys = filter(lambda x: x.startswith("tenant-"), self.db.keys())
         return [self.db.get(key) for key in tenant_keys]
 
     def get_tenant_by_name(self, tenant_name):
-        tenant_ref = self.db.get('tenant_name-%s' % tenant_name)
-        return tenant_ref
+        try:
+            return self.db.get('tenant_name-%s' % tenant_name)
+        except exception.NotFound:
+            raise exception.TenantNotFound(tenant_id=tenant_name)
 
     def get_tenant_users(self, tenant_id):
+        self.get_tenant(tenant_id)
         user_keys = filter(lambda x: x.startswith("user-"), self.db.keys())
         user_refs = [self.db.get(key) for key in user_keys]
         return filter(lambda x: tenant_id in x['tenants'], user_refs)
 
     def _get_user(self, user_id):
-        user_ref = self.db.get('user-%s' % user_id)
-        return user_ref
+        try:
+            return self.db.get('user-%s' % user_id)
+        except exception.NotFound:
+            raise exception.UserNotFound(user_id=user_id)
 
     def _get_user_by_name(self, user_name):
-        user_ref = self.db.get('user_name-%s' % user_name)
-        return user_ref
+        try:
+            return self.db.get('user_name-%s' % user_name)
+        except exception.NotFound:
+            raise exception.UserNotFound(user_id=user_name)
 
     def get_user(self, user_id):
         return _filter_user(self._get_user(user_id))
@@ -96,10 +113,16 @@ class Identity(kvs.Base, identity.Driver):
         return _filter_user(self._get_user_by_name(user_name))
 
     def get_metadata(self, user_id, tenant_id):
-        return self.db.get('metadata-%s-%s' % (tenant_id, user_id)) or {}
+        try:
+            return self.db.get('metadata-%s-%s' % (tenant_id, user_id))
+        except exception.NotFound:
+            raise exception.MetadataNotFound()
 
     def get_role(self, role_id):
-        return self.db.get('role-%s' % role_id)
+        try:
+            return self.db.get('role-%s' % role_id)
+        except exception.NotFound:
+            raise exception.RoleNotFound(role_id=role_id)
 
     def list_users(self):
         user_ids = self.db.get('user_list', [])
@@ -111,15 +134,20 @@ class Identity(kvs.Base, identity.Driver):
 
     # These should probably be part of the high-level API
     def add_user_to_tenant(self, tenant_id, user_id):
+        self.get_tenant(tenant_id)
         user_ref = self._get_user(user_id)
         tenants = set(user_ref.get('tenants', []))
         tenants.add(tenant_id)
         self.update_user(user_id, {'tenants': list(tenants)})
 
     def remove_user_from_tenant(self, tenant_id, user_id):
+        self.get_tenant(tenant_id)
         user_ref = self._get_user(user_id)
         tenants = set(user_ref.get('tenants', []))
-        tenants.remove(tenant_id)
+        try:
+            tenants.remove(tenant_id)
+        except KeyError:
+            raise exception.NotFound('User not found in tenant')
         self.update_user(user_id, {'tenants': list(tenants)})
 
     def get_tenants_for_user(self, user_id):
@@ -127,14 +155,21 @@ class Identity(kvs.Base, identity.Driver):
         return user_ref.get('tenants', [])
 
     def get_roles_for_user_and_tenant(self, user_id, tenant_id):
-        metadata_ref = self.get_metadata(user_id, tenant_id)
-        if not metadata_ref:
+        self.get_user(user_id)
+        self.get_tenant(tenant_id)
+        try:
+            metadata_ref = self.get_metadata(user_id, tenant_id)
+        except exception.MetadataNotFound:
             metadata_ref = {}
         return metadata_ref.get('roles', [])
 
     def add_role_to_user_and_tenant(self, user_id, tenant_id, role_id):
-        metadata_ref = self.get_metadata(user_id, tenant_id)
-        if not metadata_ref:
+        self.get_user(user_id)
+        self.get_tenant(tenant_id)
+        self.get_role(role_id)
+        try:
+            metadata_ref = self.get_metadata(user_id, tenant_id)
+        except exception.MetadataNotFound:
             metadata_ref = {}
         roles = set(metadata_ref.get('roles', []))
         roles.add(role_id)
@@ -142,8 +177,9 @@ class Identity(kvs.Base, identity.Driver):
         self.update_metadata(user_id, tenant_id, metadata_ref)
 
     def remove_role_from_user_and_tenant(self, user_id, tenant_id, role_id):
-        metadata_ref = self.get_metadata(user_id, tenant_id)
-        if not metadata_ref:
+        try:
+            metadata_ref = self.get_metadata(user_id, tenant_id)
+        except exception.MetadataNotFound:
             metadata_ref = {}
         roles = set(metadata_ref.get('roles', []))
         if role_id not in roles:
@@ -156,12 +192,22 @@ class Identity(kvs.Base, identity.Driver):
 
     # CRUD
     def create_user(self, user_id, user):
-        if self.get_user(user_id):
+        try:
+            self.get_user(user_id)
+        except exception.UserNotFound:
+            pass
+        else:
             msg = 'Duplicate ID, %s.' % user_id
             raise exception.Conflict(type='user', details=msg)
-        if self.get_user_by_name(user['name']):
+
+        try:
+            self.get_user_by_name(user['name'])
+        except exception.UserNotFound:
+            pass
+        else:
             msg = 'Duplicate name, %s.' % user['name']
             raise exception.Conflict(type='user', details=msg)
+
         user = _ensure_hashed_password(user)
         self.db.set('user-%s' % user_id, user)
         self.db.set('user_name-%s' % user['name'], user)
@@ -177,33 +223,49 @@ class Identity(kvs.Base, identity.Driver):
                 msg = 'Duplicate name, %s.' % user['name']
                 raise exception.Conflict(type='user', details=msg)
         # get the old name and delete it too
-        old_user = self.db.get('user-%s' % user_id)
+        try:
+            old_user = self.db.get('user-%s' % user_id)
+        except exception.NotFound:
+            raise exception.UserNotFound(user_id=user_id)
         new_user = old_user.copy()
         user = _ensure_hashed_password(user)
         new_user.update(user)
-        new_user['id'] = user_id
+        if new_user['id'] != user_id:
+            raise exception.ValidationError('Cannot change user ID')
         self.db.delete('user_name-%s' % old_user['name'])
         self.db.set('user-%s' % user_id, new_user)
         self.db.set('user_name-%s' % new_user['name'], new_user)
         return new_user
 
     def delete_user(self, user_id):
-        old_user = self.db.get('user-%s' % user_id)
+        try:
+            old_user = self.db.get('user-%s' % user_id)
+        except exception.NotFound:
+            raise exception.UserNotFound(user_id=user_id)
         self.db.delete('user_name-%s' % old_user['name'])
         self.db.delete('user-%s' % user_id)
         user_list = set(self.db.get('user_list', []))
         user_list.remove(user_id)
         self.db.set('user_list', list(user_list))
-        return None
 
     def create_tenant(self, tenant_id, tenant):
         tenant['name'] = clean.tenant_name(tenant['name'])
-        if self.get_tenant(tenant_id):
+        try:
+            self.get_tenant(tenant_id)
+        except exception.TenantNotFound:
+            pass
+        else:
             msg = 'Duplicate ID, %s.' % tenant_id
             raise exception.Conflict(type='tenant', details=msg)
-        if self.get_tenant_by_name(tenant['name']):
+
+        try:
+            self.get_tenant_by_name(tenant['name'])
+        except exception.TenantNotFound:
+            pass
+        else:
             msg = 'Duplicate name, %s.' % tenant['name']
             raise exception.Conflict(type='tenant', details=msg)
+
         self.db.set('tenant-%s' % tenant_id, tenant)
         self.db.set('tenant_name-%s' % tenant['name'], tenant)
         return tenant
@@ -211,12 +273,18 @@ class Identity(kvs.Base, identity.Driver):
     def update_tenant(self, tenant_id, tenant):
         if 'name' in tenant:
             tenant['name'] = clean.tenant_name(tenant['name'])
-            existing = self.db.get('tenant_name-%s' % tenant['name'])
-            if existing and tenant_id != existing['id']:
-                msg = 'Duplicate name, %s.' % tenant['name']
-                raise exception.Conflict(type='tenant', details=msg)
+            try:
+                existing = self.db.get('tenant_name-%s' % tenant['name'])
+                if existing and tenant_id != existing['id']:
+                    msg = 'Duplicate name, %s.' % tenant['name']
+                    raise exception.Conflict(type='tenant', details=msg)
+            except exception.NotFound:
+                pass
         # get the old name and delete it too
-        old_tenant = self.db.get('tenant-%s' % tenant_id)
+        try:
+            old_tenant = self.db.get('tenant-%s' % tenant_id)
+        except exception.NotFound:
+            raise exception.TenantNotFound(tenant_id=tenant_id)
         new_tenant = old_tenant.copy()
         new_tenant.update(tenant)
         new_tenant['id'] = tenant_id
@@ -226,10 +294,12 @@ class Identity(kvs.Base, identity.Driver):
         return new_tenant
 
     def delete_tenant(self, tenant_id):
-        old_tenant = self.db.get('tenant-%s' % tenant_id)
+        try:
+            old_tenant = self.db.get('tenant-%s' % tenant_id)
+        except exception.NotFound:
+            raise exception.TenantNotFound(tenant_id=tenant_id)
         self.db.delete('tenant_name-%s' % old_tenant['name'])
         self.db.delete('tenant-%s' % tenant_id)
-        return None
 
     def create_metadata(self, user_id, tenant_id, metadata):
         self.db.set('metadata-%s-%s' % (tenant_id, user_id), metadata)
@@ -241,15 +311,17 @@ class Identity(kvs.Base, identity.Driver):
 
     def delete_metadata(self, user_id, tenant_id):
         self.db.delete('metadata-%s-%s' % (tenant_id, user_id))
-        return None
 
     def create_role(self, role_id, role):
-        role_ref = self.get_role(role_id)
-        if role_ref:
+        try:
+            self.get_role(role_id)
+        except exception.RoleNotFound:
+            pass
+        else:
             msg = 'Duplicate ID, %s.' % role_id
             raise exception.Conflict(type='role', details=msg)
-        role_refs = self.list_roles()
-        for role_ref in role_refs:
+
+        for role_ref in self.list_roles():
             if role['name'] == role_ref['name']:
                 msg = 'Duplicate name, %s.' % role['name']
                 raise exception.Conflict(type='role', details=msg)
@@ -260,24 +332,26 @@ class Identity(kvs.Base, identity.Driver):
         return role
 
     def update_role(self, role_id, role):
-        role_refs = self.list_roles()
         old_role_ref = None
-        for role_ref in role_refs:
+        for role_ref in self.list_roles():
             if role['name'] == role_ref['name'] and role_id != role_ref['id']:
                 msg = 'Duplicate name, %s.' % role['name']
                 raise exception.Conflict(type='role', details=msg)
             if role_id == role_ref['id']:
                 old_role_ref = role_ref
-        if old_role_ref:
-            role['id'] = role_id
-            self.db.set('role-%s' % role_id, role)
-        else:
+        if old_role_ref is None:
             raise exception.RoleNotFound(role_id=role_id)
+        new_role = old_role_ref.copy()
+        new_role.update(role)
+        new_role['id'] = role_id
+        self.db.set('role-%s' % role_id, new_role)
         return role
 
     def delete_role(self, role_id):
-        self.db.delete('role-%s' % role_id)
+        try:
+            self.db.delete('role-%s' % role_id)
+        except exception.NotFound:
+            raise exception.RoleNotFound(role_id=role_id)
         role_list = set(self.db.get('role_list', []))
         role_list.remove(role_id)
         self.db.set('role_list', list(role_list))
-        return None
